Convert old tests to TS
